refactor(structure): move comments contexts to singletons
- moved CommentsIntentContext
- moved CommentsOnboardingContext
- moved CommentsSelectedPathcontext
- moved CommentsUpsellContext
