+++ packages/sanity/src/_singletons/index.ts
@@ -30,6 +30,10 @@ export * from './structure/comments/authoring-path/CommentsAuthoringPathContext'
 export * from './structure/comments/CommentsContext'
 export * from './structure/comments/components/pte/comment-input/CommentInputContext'
 export * from './structure/comments/enabled/CommentsEnabledContext'
+export * from './structure/comments/intent/CommentsIntentContext'
+export * from './structure/comments/onboarding/CommentsOnboardingContext'
+export * from './structure/comments/selected-path/CommentsSelectedPathContext'
+export * from './structure/comments/upsell/CommentsUpsellContext'
 export * from './structure/components/pane/PaneContext'
 export * from './structure/panes/document/DocumentPaneContext'
 export * from './structure/StructureToolContext'
