+++ packages/sanity/src/_singletons/structure/comments/onboarding/CommentsOnboardingContext.ts
@@ -1,5 +1,8 @@
 import {createContext} from 'react'
 
-import {type CommentsOnboardingContextValue} from './types'
+import type {CommentsOnboardingContextValue} from '../../../../structure/comments/src/context/onboarding/types'
 
+/**
+ * @internal
+ */
 export const CommentsOnboardingContext = createContext<CommentsOnboardingContextValue | null>(null)
