+++ packages/sanity/src/_singletons/structure/comments/selected-path/CommentsSelectedPathContext.ts
@@ -1,7 +1,10 @@
 import {createContext} from 'react'
 
-import {type CommentsSelectedPathContextValue} from './types'
+import type {CommentsSelectedPathContextValue} from '../../../../structure/comments/src/context/selected-path/types'
 
+/**
+ * @internal
+ */
 export const CommentsSelectedPathContext = createContext<CommentsSelectedPathContextValue | null>(
   null,
 )
