+++ packages/sanity/src/_singletons/structure/comments/upsell/CommentsUpsellContext.ts
@@ -1,9 +1,8 @@
 import {createContext} from 'react'
 
-import {type CommentsUpsellContextValue} from './types'
+import type {CommentsUpsellContextValue} from '../../../../structure/comments/src/context/upsell/types'
 
 /**
- * @beta
- * @hidden
+ * @internal
  */
 export const CommentsUpsellContext = createContext<CommentsUpsellContextValue | null>(null)
