+++ packages/sanity/src/structure/comments/src/context/intent/index.ts
@@ -1,2 +1,1 @@
-export * from './CommentsIntentContext'
 export * from './CommentsIntentProvider'
