+++ packages/sanity/src/structure/comments/src/context/selected-path/index.ts
@@ -1,3 +1,2 @@
-export * from './CommentsSelectedPathContext'
 export * from './CommentsSelectedPathProvider'
 export * from './types'
