+++ packages/sanity/src/structure/comments/src/context/upsell/index.ts
@@ -1,3 +1,2 @@
-export * from './CommentsUpsellContext'
 export * from './CommentsUpsellProvider'
 export * from './types'
